Version 1.6.45
+ card_event.py got old format in comments.
+ added webapi.config.service_mode parameter that specifies service work mode.
+ service.py
+ gevent.pywsgi got changes on line 199
  > import socket
  > self.socket._sock.shutdown(socket.SHUT_RDWR)
  This change should fix problem with pending connection to http-server in windows service mode.

+++ service.py
@@ -14,7 +14,7 @@ executable = os.path.join(os.path.dirname(sys.executable), 'u2.exe')
 multiprocessing.set_executable(executable)
 del executable
 
-BASE_PROCESS_WAIT_TIME = 2
+BASE_PROCESS_WAIT_TIME = 3
 
 class logger(object):
  def __init__(self,out):
@@ -34,33 +34,41 @@ class ServiceLauncher(win32serviceutil.ServiceFramework):
   sys.stdout = l
   sys.stderr = l
   os.chdir(os.path.dirname(sys.executable))
-  
-  # CREATE_NEW_PROCESS_GROUP with win32console.AllocConsole() are required for send_signal(signal.CTRL_BREAK_EVENT)
-  win32console.AllocConsole()
+
+  exec(open('config.py').read())
+  import webapi.config
+  self.mode = webapi.config.service_mode
+
+  servicemanager.LogInfoMsg('\nService mode = %s' % (self.mode))
+
+  if self.mode == 'process':
+   # CREATE_NEW_PROCESS_GROUP with win32console.AllocConsole() are required for send_signal(signal.CTRL_BREAK_EVENT).
+   # Both signal sender and receiver should have console allocated.
+   win32console.AllocConsole()
 
  def SvcStop(self):
-  import signal 
-
   self.ReportServiceStatus(win32service.SERVICE_STOP_PENDING)
-  self.server_process.send_signal(signal.CTRL_BREAK_EVENT)
-  time.sleep(BASE_PROCESS_WAIT_TIME)  
-  if self.server_process.poll() == None:
-   servicemanager.LogInfoMsg('\nMain process still running despite the break signal.')
-   self.server_process.kill()
-
-   #self.server.stop()
+  if self.mode == 'process':
+   import signal
+   self.server_process.send_signal(signal.CTRL_BREAK_EVENT)
+   time.sleep(BASE_PROCESS_WAIT_TIME)
+   if self.server_process.poll() == None:
+    servicemanager.LogInfoMsg('\nMain process still running despite the break signal.')
+    self.server_process.kill()
+  else:
+   self.server.stop()
 
  def SvcDoRun(self):
-  from subprocess import Popen,CREATE_NEW_PROCESS_GROUP
-  
-  self.server_process = Popen(['u2.exe','api'],creationflags = CREATE_NEW_PROCESS_GROUP)
-  self.server_process.wait()
- 
-  #exec(open('config.py').read())
+  if self.mode == 'process':
+   from subprocess import Popen,CREATE_NEW_PROCESS_GROUP
+   self.server_process = Popen(['u2.exe','api'],creationflags = CREATE_NEW_PROCESS_GROUP)
+   self.server_process.wait()
+  else:
+   exec(open('config.py').read())
 
-  #from webapi.server import make_server
-  #self.server = make_server()
-  #self.server.serve_forever()
+   from webapi.server import make_server
+   self.server = make_server()
+   self.server.serve_forever()
 
 if __name__ == '__main__':
  multiprocessing.freeze_support()
